humanizedoc/backend/requirements.txt
# HumanizeDOC — Backend Dependencies (pinned)
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
python-docx==1.1.2
pydantic==2.10.5
pydantic-settings==2.7.1
groq==0.15.0
google-generativeai==0.8.4
azure-storage-blob==12.24.1
azure-identity==1.19.0
python-dotenv==1.0.1
httpx==0.28.1
aiofiles==24.1.0
anyio==4.8.0
